backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-Mail==0.10.0
supabase==2.28.2
python-dotenv==1.2.2
PyJWT==2.12.1
gunicorn==23.0.0
